h002_customer_experience_automation/requirements.txt
streamlit==1.29.0
tornado==6.3.2
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pydantic>=2.0.0
typing-extensions>=4.5.0
tqdm>=4.65.0
faker>=19.0.0
pypdf>=3.9.0
python-multipart>=0.0.6
geopy>=2.3.0
